requirements.txt
cryptography==41.0.2
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
pip-chill==1.0.3
psycopg2-binary==2.9.6
pyjwt==2.8.0
python-dotenv==1.0.0
